Compute classic density (#42)
Add method to compute the classic density within a given measurement area. This method is available as `compute_classic_density` in the module `report.methods.density_calculator`.

The classic density is defined as N/A, where N is the number of pedestrians inside the measurement area and A is the area of the measurement area.