Add method for voronoi density calculation (#43)
Add method to compute the voronoi density within a given measurement area. This method is available as `compute_voronoi_density` in the module `report.methods.density_calculator`.

The voronoi density is formally defined as method D in `Transitions in pedestrian fundamental diagrams of straight corridors and T-junctions` by Zhang et. al (10.1088/1742-5468/2011/06/P06004). In short, create a voronoi cell for each pedestrian and calculate the proportion between the voronoi cell area and the part of the vornoi cell inside of the measurement area. The density then is the sum of the proportions divided by the area of the measurement area.