Computation of  profiles in measurement area
Add a new option to the profile computation, which allows to computation
of profiles in a specified axis-aligned measurement area. This would
allow to speed up the computation in cases where only specific areas are
of interest.

Add reference tests for the new method, which reuses the data from the
existing reference tests to verify the results. The new test cases create
a measurement area and compute the profiles for this area. This result is
compared with the to this area restricted reference results.