Add ignore_too_short_trajectories flag to load_trajectory_from_vadere
This new flag allows users to control how short trajectories are handled.
By default, ignore_too_short_trajectories is set to False. In this case, if a trajectory is too short for the specified frame rate, an error is raised.

If the flag is set to True, a warning is logged instead, and the short trajectory is ignored.