Add Vadere trajectory loader (#460)
* Add trajectory loader for Vadere trajectory (postvis.traj) files

* Refactor trajectory_loader for Vadere trajectories

* Refactor using scripts/format.sh

* Add test_load_trajectory_from_vadere_no_data

* Add test_load_trajectory_from_vadere_reference_file including reference file (created with Vadere 3.0, d79a4bfd)

* Fix unstable floating point operations in trajectory_loader

* Add test_load_trajectory_from_vadere_success

* Group vadere test functions

* Add tests for missing and non-unique columns in vadere trajectory files

* Refactor using scripts/format.sh

* Refactoring

Removed commented code

* Add load_trajectory_from_vadere to __init__.py

* Add Vadere example to jupyter notebook

* Add scenario and trajectory file

* Update user_guide.ipynb

* Add draft for load_walkable_area_from_vadere_scenario

* Update vadere scenario and trajectory files

* Add walkable area for Vadere to user guide jupyter notebook

* Simplify the way how the Vadere bounding box is loaded to PedPy

* Update jupyter notebook

* Add decimal places as input variable

This allows us to explicitly handle floating point errors that occur when transforming walkable areas and shapely polygons

* Add test_load_walkable_area_from_vadere_scenario

* Refactor error messages and syntax

* Correct input argument that leads to error

* Fix walkable area test with margin in test_trajectory_loader

Previously, the margin added to coordinates was ignored,
causing the resulting polygons to fail the equality check.

* Remove load_trajectory_from_vadere frame_rate default value

As there is no meaning of a frame rate of 24 in Vadere, we decided to remove this default value. Thus making it more transparent to the users that this needs to be passed to the function explicitly.

* Improve load_trajectory_data_from_vadere error message when required fields are missing

* Add ignore_too_short_trajectories flag to load_trajectory_from_vadere

This new flag allows users to control how short trajectories are handled.
By default, ignore_too_short_trajectories is set to False. In this case, if a trajectory is too short for the specified frame rate, an error is raised.

If the flag is set to True, a warning is logged instead, and the short trajectory is ignored.

* Remove inplace=True calls

From the ruff linter: PD002 inplace=True should be avoided; it has inconsistent behavior

* Remove default frame_rate of zero from write_vadere_csv_file

* Support overlapping obstacles when loading Vadere walkable areas

based on suggestion by @schroedtert: https://github.com/PedestrianDynamics/PedPy/pull/382#discussion_r1784816737

* Reformat Vadere trajectory code

* Improve Vadere trajectory import code quality (PR-460)

---------

Co-authored-by: Simon Rahn <simon.rahn@hm.edu>
Co-authored-by: Hayato Hess <gasi04hi@hm.edu>